backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
audioop-lts==0.2.1
av==13.1.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.36.26
botocore==1.36.26
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloup==3.0.6
colorama==0.4.6
decorator==5.2.1
Django==5.1.5
docker==7.1.0
et_xmlfile==2.0.0
fastapi==0.115.8
filelock==3.17.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
fsspec==2025.2.0
git-filter-repo==2.47.0
glcontext==3.0.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
isosurfaces==0.1.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
manim==0.19.0
ManimPango==0.6.0
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.1.1
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.2.1
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
pycairo==1.27.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyglet==2.1.3
pyglm==2.8.0
Pygments==2.19.1
PyJWT==2.10.1
pynvim==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
requests==2.32.3
rich==13.9.4
s3transfer==0.11.2
scipy==1.15.2
screeninfo==0.8.1
setuptools==75.8.2
six==1.17.0
skia-pathops==0.8.0.post2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.3
srt==3.5.3
starlette==0.45.3
supabase==2.3.0
svgelements==1.9.6
sympy==1.13.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
Werkzeug==3.1.3
